+++ pyclasher/api/models/Labels.py
@@ -1,4 +1,4 @@
-from api.models import BaseModel, IterBaseModel, IconUrls
+from .BaseModels import BaseModel, IterBaseModel, IconUrls
 
 
 class Label(BaseModel):
