+++ pyclasher/api/models/Season.py
@@ -1,4 +1,4 @@
-from Exceptions import InvalidSeasonFormat
+from ...exceptions import InvalidSeasonFormat
 
 
 class Season:
