+++ pyclasher/utils/ExecTimer.pyi
@@ -1,6 +1,3 @@
-from asyncio import Task
-
-
 class ExecutionTimer:
     def __init__(self, min_time: float = 0) -> None:
         self._min_time = min_time
