+++ tests/requests/async_tests/test_player.py
@@ -1,8 +1,8 @@
 import pytest
 
 from pyclasher import PlayerRequest, Missing
-from api.models import (ClanRole, PlayerHouse, PlayerClan, PlayerAchievementProgressList, BuilderBaseLeague,
-                        PlayerItemLevelList, LabelList, League, PlayerLegendStatistics, WarPreference)
+from pyclasher.api.models import (ClanRole, PlayerHouse, PlayerClan, PlayerAchievementProgressList, BuilderBaseLeague,
+                                  PlayerItemLevelList, LabelList, League, PlayerLegendStatistics, WarPreference)
 
 from ...constants import TEST_PLAYER_TAG
